feat(devBundler): expose css loading system for reuse
