+++ packages/one-app-dev-bundler/esbuild/plugins/styles-loader.js
@@ -14,16 +14,7 @@
  * permissions and limitations under the License.
  */
 
-import { compile as sassCompile } from 'sass';
-import postcss from 'postcss';
-import cssModules from 'postcss-modules';
-import crypto from 'crypto';
-import fs from 'node:fs';
-import cssnano from 'cssnano';
-import purgecss from '../utils/purgecss.js';
-import getModulesBundlerConfig from '../utils/get-modules-bundler-config.js';
-import { BUNDLE_TYPES } from '../constants/enums.js';
-import { addStyle } from '../utils/server-style-aggregator.js';
+import loadStyles from '../utils/load-styles.js';
 
 // the styles loader currently supports
 // scss to css conversion
@@ -32,71 +23,8 @@ import { addStyle } from '../utils/server-style-aggregator.js';
 const stylesLoader = (cssModulesOptions = {}, { bundleType } = {}) => ({
   name: 'stylesLoader',
   setup(build) {
-    const {
-      localsConvention = 'camelCaseOnly',
-      generateScopedName,
-    } = cssModulesOptions;
-
     build.onLoad({ filter: /.s?css$/ }, async (args) => {
-      // Compile scss to css
-      let cssContent;
-      if (args.path.endsWith('scss')) {
-        cssContent = sassCompile(args.path, { loadPaths: ['./node_modules'] }).css.toString();
-      } else {
-        cssContent = await fs.promises.readFile(args.path, 'utf8');
-      }
-
-      // Generate the css modules information
-      let cssModulesJSON = {};
-      const purgecssConfig = getModulesBundlerConfig('purgecss') || {};
-      const result = await postcss([
-        (process.env.NODE_ENV === 'production' && !purgecssConfig.disabled) && purgecss(purgecssConfig),
-        cssModules({
-          localsConvention,
-          generateScopedName,
-          getJSON(_cssSourceFile, json) {
-            cssModulesJSON = { ...json };
-            return cssModulesJSON;
-          },
-        }),
-        process.env.NODE_ENV === 'production' && cssnano({
-          preset: 'default',
-        }),
-      ].filter(Boolean)).process(cssContent, {
-        from: undefined,
-        map: false,
-      });
-
-      const hash = crypto.createHash('sha256');
-      hash.update(result.css);
-      const digest = hash.copy().digest('hex');
-
-      let injectedCode = '';
-      if (bundleType === BUNDLE_TYPES.BROWSER) {
-        // For browsers generate code to inject this style into the head at runtime
-        injectedCode = `\
-(function() {
-  if ( global.BROWSER && !document.getElementById(digest)) {
-    var el = document.createElement('style');
-    el.id = digest;
-    el.textContent = css;
-    document.head.appendChild(el);
-  }
-})();`;
-      } else {
-        // For SSR, aggregate all styles, then inject them once at the end
-        const isDependencyFile = args.path.indexOf('/node_modules/') >= 0;
-        addStyle(digest, result.css, isDependencyFile);
-      }
-
-      // provide useful values to the importer of this file, most importantly, the classnames
-      const jsContent = `\
-const digest = '${digest}';
-const css = \`${result.css}\`;
-${injectedCode}
-${Object.entries(cssModulesJSON).map(([exportName, className]) => `export const ${exportName} = '${className}';`).join('\n')}
-export default { ${Object.keys(cssModulesJSON).join(', ')} };
-export { css, digest };`;
+      const jsContent = await loadStyles({ path: args.path, cssModulesOptions, bundleType });
 
       return {
         contents: jsContent,
